@@ -1,9 +1,9 @@
 # This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
+# The data can then be loaded with the rails db:seed (or created alongside the db with db:setup).
 #
 # Examples:
 #
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
-user1 = User.create :name => "User1", :email => "user1@example.com", :password => "password"
+user1 = User.create :name => "CodeceptJS Tester", :email => "codeceptjs-tester@example.com", :password => "password"
