@@ -1,7 +1,7 @@
 
-Feature('Basic Homepage');
+Feature('Basic Homepage Test');
 
-Scenario('test homepage', (I) => {
+Scenario('test homepage loads', (I) => {
 	I.amOnPage('/');
 	I.seeElement('#home-sl1'); // check for not signed in homepage element
 });
